[front/bid]wip: socket send bid price
