+++ front/views/home/home.html
@@ -248,7 +248,7 @@
           <button class="like_btn" type="submit">
             <i class="fa-solid fa-heart"> 찜하기</i>
           </button>
-          <button id="contact_btn" type="submit">1:1 연락</button>
+          <button class="contact_btn" type="submit">1:1 연락</button>
         </div>
       </article>
     </section>
